feat: Add attribute to calibration graphs as needsCalibration type and another as calibrationMode.
What this helps to know the code that the calibrationMode graph is where calibration points can be selected and needscalibration graph is where the actual wavelength value is reflected. Also this change includes additional text label in needsCalibration graph. This indicates if the calibrating graph has finised calibration or not.

 Please enter the commit message for your changes. Lines starting